Remove the RBX tests since they fail on about 1/2 the platforms
rbx seems to be nearly impossible to install reliably. People are going to have to do some serious attribute / PPA magic to get this thing to work. For the tests that's just not possible and it keeps us from having reliable tests that can run on every Github PR. It's worth dumping to get that.

Signed-off-by: Tim Smith <tsmith@chef.io>